+++ CHANGELOG.md
@@ -63,6 +63,7 @@
         - selenium-webdriver: 4.1.1 to 4.3.1
         - sinon: 12.0.1 to 14.0.0
         - zip-a-folder: 1.1.3 to 1.1.5
+- WfsSearch: `inputLabel` are now translated.
 
 ### Deprecated